+++ pkg/costco/constants.go
@@ -2,7 +2,7 @@ package costco
 
 // Library Version
 const (
-	Version = "0.3.12"
+	Version = "0.3.13"
 )
 
 // API Endpoints